flink version from 1.10.1 to 1.13.0; add rabbitmq source&sink Demo; 增加mysql ddl流读取写入

+++ kudu/src/main/java/com/xh/kudu/service/KuduServiceImpl.java
@@ -48,7 +48,10 @@ public class KuduServiceImpl implements KuduService {
             List<String> tblist = listTablesResponse.getTablesList();
             for(String tableName : tblist) {
                 System.out.println(tableName);
-                client.deleteTable(tableName);
+
+                if(tableName.startsWith("impala::kudu_"+dbKey)){
+                    client.deleteTable(tableName);
+                }
             }
         }catch (Exception e){
             e.printStackTrace();
